+++ app/api/utils/rapid.py
@@ -1,6 +1,6 @@
 from loguru import logger
 import requests
-from core.config import NCBI_DATASETS_URL
+from core.config import ENSEMBL_URL, NCBI_DATASETS_URL
 import re
 
 
@@ -52,3 +52,26 @@ def format_assembly_accession(species_url_name: str):
         return assembly_accession_id
     else:
         return None
+
+
+def construct_url(genome_id, subpath, query_params):
+    location = query_params.get("r", [None])[0]
+    gene_id = query_params.get("g", [None])[0]
+
+    if subpath == "" or re.search("Info/Index", subpath, re.IGNORECASE):
+        return f"{ENSEMBL_URL}/species/{genome_id}"
+    elif re.search("Location", subpath):
+        return f"{ENSEMBL_URL}/genome-browser/{genome_id}?focus=location:{location}"
+    elif re.search("Gene", subpath):
+        if re.search("Gene/Compara_Homolog", subpath):
+            return (
+                f"{ENSEMBL_URL}/entity-viewer/{genome_id}/gene:{gene_id}?view=homology"
+            )
+        return f"{ENSEMBL_URL}/entity-viewer/{genome_id}/gene:{gene_id}"
+    elif re.search("Transcript", subpath):
+        if re.search("Domains|ProteinSummary", subpath):
+            return (
+                f"{ENSEMBL_URL}/entity-viewer/{genome_id}/gene:{gene_id}?view=protein"
+            )
+        return f"{ENSEMBL_URL}/entity-viewer/{genome_id}/gene:{gene_id}"
+    return ENSEMBL_URL
